requirements.txt
pandas
pyarrow
fastparquet
requests
sqlalchemy
psycopg2-binary
python-dotenv
